+++ webpack.common.js
@@ -1,5 +1,7 @@
-const HtmlWebpackPlugin = require("html-webpack-plugin");
 const path = require("path");
+
+const HtmlWebpackPlugin = require("html-webpack-plugin");
+const CopyPlugin = require("copy-webpack-plugin");
 
 module.exports = {
   entry: {
@@ -40,6 +42,14 @@ module.exports = {
       title: "webxdc-dev",
       template: "./frontend/index.html",
       chunks: ["frontend"],
+    }),
+    new CopyPlugin({
+      patterns: [
+        {
+          from: path.resolve(__dirname, "frontend/favicon.ico"),
+          to: path.resolve(__dirname, "dist"),
+        },
+      ],
     }),
   ],
 };