+++ CHANGELOG.md
@@ -7,6 +7,10 @@ release date when you use `npm version` (see `README.md`).
 
 ## [Unreleased]
 
+## Added
+
+- Show version number of dev tool in info tab
+
 ## [0.10.0][] - 2022-06-28
 
 ### Added
@@ -201,7 +205,5 @@ release date when you use `npm version` (see `README.md`).
 [0.8.0]: https://github.com/webxdc/webxdc-dev/tree/v0.8.0
 [unreleased]: https://github.com/webxdc/webxdc-dev/compare/v0.9.0...HEAD
 [0.9.0]: https://github.com/webxdc/webxdc-dev/tree/v0.9.0
-
-
-[Unreleased]: https://github.com/webxdc/webxdc-dev/compare/v0.10.0...HEAD
+[unreleased]: https://github.com/webxdc/webxdc-dev/compare/v0.10.0...HEAD
 [0.10.0]: https://github.com/webxdc/webxdc-dev/tree/v0.10.0